Rename logger into Logger
